refactor: map을 menu info map으로 변경하기
